requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
agate==1.9.1
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.3
daff==1.4.2
db-dtypes==1.4.3
debugpy==1.8.17
decorator==5.2.1
deepdiff==8.6.1
defusedxml==0.7.1
docstring_parser==0.17.0
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
fqdn==1.5.1
google-api-core==2.28.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-cloud-aiplatform==1.126.1
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-dataproc==5.23.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-genai==1.49.0
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
lark==1.3.1
leather==0.4.0
MarkupSafe==3.0.3
mashumaro==3.14
matplotlib-inline==0.2.1
mistune==3.1.4
more-itertools==10.8.0
msgpack==1.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.3.4
oauthlib==3.3.1
openpyxl==3.1.5
orderly-set==5.5.0
packaging==25.0
pandas==2.3.3
pandas-gbq==0.30.0
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydata-google-auth==1.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.28.0
rsa==4.9.1
Send2Trash==1.8.3
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.15
zipp==3.23.0
astronomer-cosmos[dbt-bigquery]
apache-airflow-providers-google